pyproject.toml
[project]
name = "stencil-ui"
version = "0.2.6"
description = "CLI tool to generate UI code for HTML, ImGui, or Curses from a simple config file"
readme = "README.md"
authors = [{name="Krishanth K"}]
license = {file="LICENSE"}
requires-python = ">=3.9"

dependencies = [
    "PyYAML>=6.0.3",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
imgui = [
    "glfw",
    "PyOpenGL",
    "imgui[glfw]"
]

[project.scripts]
stencil = "stencil.cli:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["stencil*"]
exclude = ["my_html_app*", "my_flutter_app*"]
